fast_master/car_api/pyproject.toml
[project]
name = "car-api"
version = "0.1.0"
description = ""
authors = [
    {name = "lphillipe",email = "luisphillipe200@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13, <4.0"
dependencies = [
    "fastapi[standard] (>=0.128.0,<0.129.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "sqlalchemy[asyncio] (>=2.0.46,<3.0.0)",
    "aiosqlite (>=0.22.1,<0.23.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "alembic (>=1.18.3,<2.0.0)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)",
    "pyjwt (>=2.11.0,<3.0.0)",
    "mkdocs (>=1.6.1,<2.0.0)",
    "mkdocs-material (>=9.7.1,<10.0.0)",
    "pymdown-extensions (>=10.20.1,<11.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.15.1,<0.16.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)"
]

[tool.ruff]
line-length = 79
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".gitignore",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "__pycache__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "alembic",
    "migrations",
]

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']
ignore = ['PLR2004', 'PLR0917', 'PLR0913']

[tool.ruff.lint.per-file-ignores]
'tests/*.py' = ['PLR6301']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
asyncio_default_fixture_loop_scope = 'function'

[tool.coverage.run]
concurrency = ["thread", "greenlet"]

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev car_api/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=car_api -vv'
post_test = 'coverage html'
docs = 'mkdocs serve -a 127.0.0.1:8001'